Switch from double to float

+++ src/accel/AABB.h
@@ -17,16 +17,16 @@ struct AABB {
     AABB() = default;
     AABB(const Vec3& minCorner, const Vec3& maxCorner) : min(minCorner), max(maxCorner) {}
 
-    inline bool hit(const Ray& ray, double tMin, double tMax) const {
+    inline bool hit(const Ray& ray, float tMin, float tMax) const {
         ++gAABBHits;
-        constexpr double EPS = 1e-8;
+        constexpr float EPS = 1e-8;
 
         for (int axis = 0; axis < 3; ++axis) {
-            double dir = ray.direction[axis];
-            double t0, t1;
+            float dir = ray.direction[axis];
+            float t0, t1;
 
             if (std::abs(dir) > EPS) { // Treat nearly parallel rays as parellel to avoid numerical blowup 
-                double invDir = 1.0 / dir;
+                float invDir = 1.0 / dir;
                 t0 = (min[axis] - ray.origin[axis]) * invDir;
                 t1 = (max[axis] - ray.origin[axis]) * invDir;
                 if (invDir < 0) std::swap(t0, t1);
