+++ src/materials/Dielectric.cpp
@@ -5,7 +5,7 @@
 #include <cmath>
 #include <random>
 
-Dielectric::Dielectric(double eta) : eta(eta) {}
+Dielectric::Dielectric(float eta) : eta(eta) {}
 
 bool Dielectric::scatter(
     const HitRecord& record,
@@ -14,7 +14,7 @@ bool Dielectric::scatter(
     Vec3& attenuation
 ) const {
     static std::mt19937 gen(std::random_device{}());
-    static std::uniform_real_distribution<double> dis(0.0, 1.0);
+    static std::uniform_real_distribution<float> dis(0.0, 1.0);
 
     // Dielectrics do not absorb light
     attenuation = Vec3(1.0, 1.0, 1.0);
@@ -25,10 +25,10 @@ bool Dielectric::scatter(
     // Determine refraction ratio depending on ray direction
     // If hitting front face: air -> material
     // If inside material: material -> air
-    double etaRatio = record.frontFace ? (1.0 / eta) : eta;
+    float etaRatio = record.frontFace ? (1.0 / eta) : eta;
 
     // Cosine of angle between ray and surface normal
-    double cosTheta = std::min(dot(-unitDir, record.normal), 1.0);
+    float cosTheta = std::min(dot(-unitDir, record.normal), 1.0f);
 
     // Compute perpendicular component of refracted ray
     Vec3 r_perp = etaRatio * (unitDir + cosTheta * record.normal);
@@ -37,9 +37,9 @@ bool Dielectric::scatter(
     bool cannotRefract = r_perp.lengthSquared() > 1.0;
 
     // Schlick approximation for reflectance
-    double R0 = (1.0 - eta) / (1.0 + eta);
+    float R0 = (1.0 - eta) / (1.0 + eta);
     R0 = R0 * R0;
-    double R = R0 + (1.0 - R0) * std::pow(1.0 - cosTheta, 5.0);
+    float R = R0 + (1.0 - R0) * std::pow(1.0 - cosTheta, 5.0);
 
     if (cannotRefract || dis(gen) < R) {
         // Reflect
