+++ src/materials/Dielectric.h
@@ -5,9 +5,9 @@
  * Transparent material that refracts and refrects light
  */
 struct Dielectric : public Material {
-    double eta; // Index of refraction: how much light bends (air=1.0, water=1.33, glass=1.5, diamond=2.4)
+    float eta; // Index of refraction: how much light bends (air=1.0, water=1.33, glass=1.5, diamond=2.4)
 
-    Dielectric(double eta);
+    Dielectric(float eta);
 
     bool scatter(
         const HitRecord& record,
